Extract shared test-connection logic into lib/db/test-connection.ts
Deduplicate TestConnectionResult type, testDatabaseConnection,
handleDatabaseTest, and handlePluginTest from both test route files
into a single shared module.
